Implemented H- bound-free and free-free opacity model as a stand-alone modeling object. Documented and tested.
In pyrat.set_spectrum() fixed bug that was turning on ion_opacity
even if there were no ions in the atmosphere.
(could be a prettier fix in theory ...)

Refactored molmodel, molfree, and molpars arguments to initialize
by default as zero-length iterables instead of None.
This fixes an error when a user launches a run with tea chemistry
by no modeling paramters.